Fix the following:   - Validation tests after changing validation limit for board sizes to 20 for 'Unique' and 'All' board sizes   - No. of Solutions type from int to ulong through the entire solution

+++ NQueen.Domain/Utils/SymmetryHelper.cs
@@ -257,7 +257,7 @@ public static partial class SymmetryHelper
     // --- Additional helpers (retained) ---
     /// <summary>
     /// Builds a title for solution listings with truncation / mode awareness.</summary>
-    public static string SolutionTitle(SolutionMode solutionMode, int noOfSolutions)
+    public static string SolutionTitle(SolutionMode solutionMode, ulong noOfSolutions)
     {
         if (solutionMode == SolutionMode.Single)
         { return "Solution:"; }
